DRAW-94 refactor: token 비즈니스 로직 -> core 모듈 이동
