+++ app/support/auth/build.gradle.kts
@@ -6,7 +6,6 @@ plugins {
 
 dependencies {
     implementation(project(":core"))
-    implementation(project(":support:jwt"))
 
     implementation("org.springframework.boot:spring-boot-starter-web:${Versions.SPRING_BOOT}")
     implementation("org.springframework.boot:spring-boot-starter-security:${Versions.SPRING_BOOT}")
