+++ core/build.gradle.kts
@@ -9,6 +9,7 @@ dependencies {
     implementation(project(":adapter:memory"))
     implementation(project(":adapter:oauth"))
     implementation(project(":adapter:rdb"))
+    implementation(project(":support:jwt"))
 
     implementation("org.springframework.boot:spring-boot-starter:${Versions.SPRING_BOOT}")
     compileOnly("org.springframework.boot:spring-boot-starter-jdbc:${Versions.SPRING_BOOT}")
